jarvis-registry refactoring the current legacy code
